fix: change type publication year to uint16

+++ internal/domain/book.go
@@ -6,5 +6,5 @@ type Book struct {
 	ID              uuid.UUID `json:"id"`
 	Title           string    `json:"title,omitempty"`
 	Author          string    `json:"author,omitempty"`
-	PublicationYear int8      `json:"publication_year,omitempty"`
+	PublicationYear uint16    `json:"publication_year,omitempty"`
 }
